requirements.txt
PyGithub
python-dotenv
discord.py
